refactor: reformulando a estrutura de pastas
